Build break fix
Summary:
Fixing build break on Mac

(1) uint64_t fix
(2) O_DIRECT works only for Linux

Test Plan: Build and test on Mac and Unix

Subscribers: andrewkr, dhruba, leveldb

Differential Revision: https://reviews.facebook.net/D61353

+++ utilities/persistent_cache/block_cache_tier.cc
@@ -338,7 +338,7 @@ bool BlockCacheTier::Reserve(const size_t size) {
       return false;
     }
     assert(!f->refs_);
-    size_t file_size;
+    uint64_t file_size;
     if (!f->Delete(&file_size).ok()) {
       // unable to delete file
       return false;
